[REFACTOR] Remove use of pages directory, moving things to routes instead, also refactor things to flat routes
