+++ vitest.config.ts
@@ -19,7 +19,7 @@ export default mergeConfig(
           '**/*.{spec,test}.{ts,tsx}',
           '{postcss,tailwind,eslint,vite,vitest}.{config,setup}.{js,ts}',
           // Ignore chart from tests, since it uses a third part library anyway
-          './src/components/atoms/chart.tsx',
+          'src/components/atoms/chart.tsx',
           'src/vite-env.d.ts',
           'src/main.tsx',
           // TODO: ignore schema for now
